requirements.txt
praw
google-generativeai
python-dotenv
torch
transformers